ai-studio/backend/requirements.txt
# FastAPI Backend Server Operations
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4

# API Routing & Networking
requests==2.31.0
httpx==0.27.0
tenacity==8.2.3

# Agent Orchestration (LangGraph + LangChain)
langgraph==0.0.31
langchain-core==0.1.40
langchain-groq==0.1.2

# Database Connectivity
supabase==2.4.3

# Phase 3 — Image Generation
pillow==10.3.0

# ─────────────────────────────────────────────
# Phase 4: Scaling & Observability
# ─────────────────────────────────────────────

# Structured JSON Logging (replaces print statements)
structlog==24.1.0

# Rate Limiting middleware
slowapi==0.1.9

# Redis-backed ARQ Worker Queue (crash-safe background jobs)
redis==5.0.4
arq==0.25.0

# asyncpg connection pooling (Supabase Postgres direct)
asyncpg==0.29.0

# ─────────────────────────────────────────────
# Phase 4: Testing Infrastructure
# ─────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0  # Used by TestClient
elevenlabs==1.0.0
edge-tts==6.1.10
moviepy==1.0.3
gradio_client==1.3.0
numpy==1.26.4
